Renamed backend to core
